Fix libretro core name
== DETAILS
When the APPNAME macro was added, the libretro code was included
in the change and doing so broke netplay in RetroArch.

RetroArch stores core information in *.info files, and the `corename`
field of the info file is expected to match what the core self-reports.
The result of the APPNAME macro work is that it changed the self-reporting
core name as "mame2003".

This commit changes the core self-reported name (and JUST that name) back
to the expected value.

== TESTING
Tested locally and confirmed that the content matching in netplay is now
successful.